Expand variables in paths early
Variables can be used in paths (`${workspaceFolder}` or `${env:HOME}`)
but they were already being resolved to absolute paths before expanding
the variables, which made the expansions not do the "right" thing.

This commit changes it to expand the variables early, in the server
constructor.
